add sentinel value to terminate the SUPPORTED_OBJECT_TABLE
Because of how the bacnet-stack works, it is recommended to terminate the `SUPPORTED_OBJECT_TABLE` with a sentinel value that indicates the end.

Steve Karg's usual way is to set the last value of the list to a value completely null with the exception of the type in which is set `MAX_BACNET_OBJECT_TYPE`.

Co-authored-by: Ed <edward@bac-test.com>